+++ spec/lib/karafka/cli/console_spec.rb
@@ -1,20 +1,23 @@
 require 'spec_helper'
 
-RSpec.describe Karafka::Cli do
-  subject { described_class.new }
+RSpec.describe Karafka::Cli::Console do
+  let(:cli) { Karafka::Cli.new }
+  subject { described_class.new(cli) }
 
-  describe '#console' do
+  specify { expect(described_class).to be < Karafka::Cli::Base }
+
+  describe '#call' do
     let(:cmd) { "KARAFKA_CONSOLE=true bundle exec irb -r #{Karafka.boot_file}" }
 
     it 'expect to execute irb with boot file required' do
-      expect(subject)
+      expect(cli)
         .to receive(:info)
 
       expect(subject)
         .to receive(:system)
         .with(cmd)
 
-      subject.console
+      subject.call
     end
   end
 end